Move handling of base_font_size to canvas
